Build fixes for FreeBSD 10

+++ src/terminal/terminal.cc
@@ -45,7 +45,7 @@ Emulator::Emulator( size_t s_width, size_t s_height )
   : fb( s_width, s_height ), dispatch(), user()
 {
   /* explode early */
-  assert(MB_CUR_MAX < mb_cur_max);
+  assert((size_t)MB_CUR_MAX < mb_cur_max);
 }
 
 std::string Emulator::read_octets_to_host( void )
